requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
openai==1.3.7
python-dotenv==1.0.0
lxml==4.9.3
selenium==4.32.0
supabase==2.17.0
webdriver-manager==4.0.2